Add TableScan constructor
